Fix clippy warnings and improve CI
Clippy fixes:
- Use strip_prefix() instead of manual string slicing
- Replace while let with for loop
- Remove unnecessary usize casts
- Derive Default for CompressionFormat
- Use clamp() instead of max().min()
- Simplify UTF-8 byte range match

CI improvements:
- Dynamic core detection: benchmark with 1 thread and max cores
- Create unrigz symlink before validation
- Install system gzip for baseline comparisons
- Add opt-level=3 and strip=true to release profile

+++ src/compression.rs
@@ -67,7 +67,7 @@ pub fn compress_file(filename: &str, args: &RigzArgs) -> RigzResult<i32> {
 
     // Create optimization configuration
     let opt_config = OptimizationConfig::new(
-        args.processes as usize,
+        args.processes,
         file_size,
         args.compression_level,
         content_type,
@@ -172,7 +172,7 @@ pub fn compress_stdin(args: &RigzArgs) -> RigzResult<i32> {
 
     // Create optimization configuration
     let opt_config = OptimizationConfig::new(
-        args.processes as usize,
+        args.processes,
         file_size,
         args.compression_level,
         content_type,
